orchestrator/requirements.txt
google-genai>=0.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
google-cloud-pubsub>=2.23.0
google-cloud-secret-manager>=2.20.0
supabase>=2.9.0
redis>=5.0.0
firebase-admin>=6.5.0
python-multipart>=0.0.9
websockets>=13.0
pydantic>=2.9.0
httpx>=0.27.0
